-Commonize the duplicate idplayer() and validplayer() functions from the client and the server versions of g_game.cpp and put the common functions into p_user.cpp with other player related functions. -Cache all ids while searching for a player in idplayer() rather than just the id of the player we're interested in. -Remove a couple warnings from p_user.cpp
SVN r2647 (0.6-dev)

+++ common/d_player.h
@@ -529,7 +529,7 @@ extern std::vector<player_t> players;
 player_t		&consoleplayer();
 player_t		&displayplayer();
 player_t		&listenplayer();
-player_t		&idplayer(size_t id);
+player_t		&idplayer(byte id);
 bool			validplayer(player_t &ref);
 
 extern byte consoleplayer_id;
